+++ src/Oro/Component/Action/Exception/InvalidParameterException.php
@@ -2,6 +2,6 @@
 
 namespace Oro\Component\Action\Exception;
 
-class InvalidParameterException extends \Exception
+class InvalidParameterException extends InvalidArgumentException
 {
 }